Refactor public variables to private in UI
